#67 - law instance 데이터 처리 step 작업 - sample 테스트 완료
#67 - law instance 데이터 처리 step 작업

+++ spring-mermer-batch/src/main/java/com/mermer/mermerbatch/job/DomainJobConfig.java
@@ -154,8 +154,8 @@ public class DomainJobConfig {
 		return items -> {
 			items.forEach(item -> {
 				Domain domain = Domain.builder()
-								.lawId(Long.parseLong(item.getLawId()))
-								.lawMST(Long.parseLong(item.getLawSerial()))
+								.lawId(Integer.parseInt(item.getLawId()))
+								.lawMST(Integer.parseInt(item.getLawSerial()))
 								.lawName(item.getLawName())
 								.inster(99999999L)
 								.mdfer(99999999L)
